requirements.txt
fastapi
uvicorn
python-telegram-bot
SQLAlchemy
PyMySQL
cryptography
python-dotenv
openai
PyMuPDF
python-multipart
pandas
openpyxl
